% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rice.R
\name{check_token}
\alias{check_token}
\title{Check the Token}
\usage{
check_token(token)
}
\arguments{
\item{token}{a character vector of length 1 representing the outcomes of 5 rolls of a dice.}
}
\value{
\code{TRUE} if correctly formatted, \code{FALSE} otherwise.
}
\description{
Checks the validity of the token.
}
\details{
The token is a 5-digit number representing the outcomes of 5 rolls
of a 6-faced dice. The function returns \code{TRUE} if the
\code{token} is correctly formatted and \code{FALSE} otherwise.
}
\author{
Francois Michonneau
}

