% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rice.R
\name{generate_token}
\alias{generate_token}
\title{Generate the tokens}
\usage{
generate_token(n_words)
}
\arguments{
\item{n_words}{The number of token to generate}
}
\value{
A character vector representing the generated tokens.
}
\description{
Generates the tokens.
}
\details{
This function generates as many tokens as needed to create the passphrase.
}
\author{
Francois Michonneau
}

