% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rice.R
\name{match_token}
\alias{match_token}
\title{Match the token to a word}
\usage{
match_token(token, wordlist = riceware::wordlist_en, title_case = TRUE)
}
\arguments{
\item{token}{A correctly formatted token}

\item{wordlist}{A \code{data.frame} with at least 2 columns: one
named \sQuote{token} that contains the tokens, one named
\sQuote{word} that contains the words corresponding to the tokens.}

\item{title_case}{If \code{TRUE}, the first letter of each word in
the passphrase will be capitalized}
}
\value{
A character vector of length 1 representing the word
corresponding to the token.
}
\description{
Retrieves the word corresponding to a given token.
}
\details{
Given a token and a list of words, this function returns the word
matching the supplied token.
}
\author{
Francois Michonneau
}

