% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordlist.R
\docType{data}
\name{wordlist_de}
\alias{wordlist_de}
\title{Diceware word list (German)}
\format{A \code{data.frame} with  2 columns: \code{token} a 5-digit
number representing the results of 5 rolls of a dice, \code{word}
the Diceware word corresponding to the dice rolls.}
\usage{
wordlist_de
}
\description{
The German Diceware word list as provided at
\url{http://world.std.com/~reinhold/diceware_german.txt}. This
list is provided by Benjamin Tenne under the terms of the GNU
General Public License.
}
\keyword{datasets}

