% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordlist.R
\docType{data}
\name{wordlist_en}
\alias{wordlist_en}
\title{Diceware word list (English)}
\format{A \code{data.frame} with 2 columns: \code{token} a 5-digit
number representing the results of 5 rolls of a dice, \code{word}
the Diceware word corresponding to the dice rolls.}
\usage{
wordlist_en
}
\description{
The English Diceware word list as provided at
\url{http://world.std.com/~reinhold/diceware.wordlist.asc}.  To
the extent that a word list is protected by copyright, A G
Reinhold licenses its rights to the English Diceware Wordlist
under the Creative Commons CC-BY 3.0 license. Diceware is a
trademark of Arnold G. Reinhold.
}
\keyword{datasets}

