% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordlist.R
\docType{data}
\name{wordlist_es}
\alias{wordlist_es}
\title{Diceware word list (Spanish)}
\format{A \code{data.frame} with 2 columns: \code{token} a 5-digit
number representing the results of 5 rolls of a dice, \code{word}
the Diceware word corresponding to the dice rolls.}
\usage{
wordlist_es
}
\description{
The Spanish Diceware word list as provided at
\url{http://world.std.com/~reinhold/diceware_espanol/DW-Espanol-1.txt}. This
list is provided by Manuel Palao, CISM, CISA (manuel@palao.com) under the terms of the
GNU General Public License.
}
\keyword{datasets}

