% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordlist.R
\docType{data}
\name{wordlist_fr}
\alias{wordlist_fr}
\title{Diceware word list (French)}
\format{A \code{data.frame} with 2 columns: \code{token} a 5-digit
number representing the results of 5 rolls of a dice, \code{word}
the Diceware word corresponding to the dice rolls.}
\usage{
wordlist_fr
}
\description{
The French Diceware list as provided at
\url{http://weber.fi.eu.org/software/diceware/src/francais.wordlist.asc}. This list is provided
by Matthieu Weber, no licensing information is provided.
}
\keyword{datasets}

