% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordlist.R
\docType{data}
\name{wordlist_it}
\alias{wordlist_it}
\title{Diceware word list (Italian)}
\format{A \code{data.frame} with 2 columns: \code{token} a 5-digit
number representing the results of 5 rolls of a dice, \code{word}
the Diceware word corresponding to the dice rolls.}
\usage{
wordlist_it
}
\description{
The Italian Diceware list as provided at
\url{http://www.taringamberini.com/download/diceware_it_IT/word_list_diceware_in_italiano.txt}.
This list is provided by Tarin Gamberini under a GNU General Public License.
}
\keyword{datasets}

