% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordlist.R
\docType{data}
\name{wordlist_jp}
\alias{wordlist_jp}
\title{Diceware word list (Japanese)}
\format{A \code{data.frame} with 2 columns: \code{token} a 5-digit
number representing the results of 5 rolls of a dice, \code{word}
the Diceware word corresponding to the dice rolls.}
\usage{
wordlist_jp
}
\description{
The Japanese Diceware list as provided at
\url{http://s3.amazonaws.com/dotclue.org/diceware_jp.txt}
translated by Hiroshi Yuki, and provided in Romaji by J Greely
}
\keyword{datasets}

