% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordlist.R
\docType{data}
\name{wordlist_sv}
\alias{wordlist_sv}
\title{Diceware word list (Swedish)}
\format{A \code{data.frame} with 2 columns: \code{token} a 5-digit
number representing the results of 5 rolls of a dice, \code{word}
the Diceware word corresponding to the dice rolls.}
\usage{
wordlist_sv
}
\description{
The Swedish Diceware list as provided at
\url{http://x42.com/diceware/diceware-sv.txt} provided by by
Magnus Bodin. No license information provided.
}
\keyword{datasets}

